C#: Only extract a single empty location.
