+++ csharp/extractor/Semmle.Extraction.CSharp/Extractor/Context.cs
@@ -582,14 +582,14 @@ namespace Semmle.Extraction.CSharp
         public Entities.Location CreateLocation()
         {
             return SourceTree is null
-                ? Entities.GeneratedLocation.Create(this)
+                ? Entities.EmptyLocation.Create(this)
                 : CreateLocation(Microsoft.CodeAnalysis.Location.Create(SourceTree, Microsoft.CodeAnalysis.Text.TextSpan.FromBounds(0, 0)));
         }
 
         public Entities.Location CreateLocation(Microsoft.CodeAnalysis.Location? location)
         {
             return (location is null || location.Kind == LocationKind.None)
-                ? Entities.GeneratedLocation.Create(this)
+                ? Entities.EmptyLocation.Create(this)
                 : location.IsInSource
                     ? Entities.NonGeneratedSourceLocation.Create(this, location)
                     : Entities.Assembly.Create(this, location);
